Histogram Equalization
Histogram equalization is used to enhance contrast. It is not necessary that contrast will always be increase in this. There may be some cases were histogram equalization can be worse. In that cases the contrast is decreased. Histogram equalization is an image processing method to adjust the contrast of an image using its intensity distribution histogram. The equalization method increases the value of global contrast when the close contrast values represent the usable data for the image.